Refactor: Organize Zustand store into multiple slices
This commit refactors the main Zustand store (`useAppStore.js`, now `rootStore.js`) into multiple, more focused slices to improve your code's organization and maintainability.

The following slices have been created under `src/store/`:
- `screenStore.js`: Manages screen-related state and actions.
- `componentStore.js`: Manages component-related state (like IDs, selection) and actions for component manipulation (add, delete, move, resize, etc.).
- `canvasStore.js`: Manages canvas/editor view state (zoom, pan) and related actions.
- `interactionsStore.js`: Manages UI interaction states (like drag-and-drop feedback, panning state) and related actions.
- `ioStore.js`: Manages import/export logic.

The `rootStore.js` now imports these individual slice creators and combines them using Zustand's `create((set, get) => ({ ...slice1(set, get), ...slice2(set, get) }))` pattern.

Utility functions like `getAbsolutePosition` have been moved to `src/utils/positionUtils.js`.

All imports throughout the application have been updated or verified to point to the correct store or utility. Component access to store state and actions remains consistent due to the way slices are combined.

My analysis indicates all existing functionalities (screen and component management, canvas interactions, import/export) are preserved and working as expected.
